Backend/stripe/requirements.txt
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.3.0
Flask-Cors==3.0.10
requests==2.31.0
stripe==8.4.0
python-dateutil==2.8.2
python-dotenv==0.21.0